requirements.txt
mss>=9.0.0
pyautogui>=0.9.54
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
gymnasium>=0.29.0
stable-baselines3>=2.2.0
